pyproject.toml
[project]
name = "wisper-sample"
version = "0.1.0"
description = ""
authors = [
    {name = "Noriaki Hamamoto",email = "norisuke3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyaudio (>=0.2.14,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai-whisper @ git+https://github.com/openai/whisper.git"
]

[tool.poetry]
packages = [{include = "wisper_sample", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
